EVEREST-394-resources-step-unclear-values
