+++ apps/everest/src/pages/database-form/steps/resources/resources-step.tsx
@@ -11,7 +11,10 @@ import { ResourceInput } from './resource-input/resource-input';
 import { DEFAULT_SIZES } from './resources-step.const';
 import { Messages } from './resources-step.messages.ts';
 import { ResourceSize } from './resources-step.types';
-import { humanizeResourceSizeMap } from './resources-step.utils.ts';
+import {
+  checkResourceText,
+  humanizeResourceSizeMap,
+} from './resources-step.utils.ts';
 import { NODES_DB_TYPE_MAP } from '../../database-form.constants';
 import { StepHeader } from '../step-header/step-header.tsx';
 
@@ -22,35 +25,6 @@ export const ResourcesStep = () => {
     useKubernetesClusterResourcesInfo();
 
   const { isMobile, isDesktop } = useActiveBreakpoint();
-  const checkResourceText = (
-    value: string | number | undefined,
-    units: string,
-    fieldLabel: string,
-    exceedFlag: boolean
-  ) => {
-    if (value) {
-      const parsedNumber = Number(value);
-
-      if (Number.isNaN(parsedNumber)) {
-        return '';
-      }
-
-      const processedValue =
-        fieldLabel === Messages.labels.cpu
-          ? parsedNumber / 1000
-          : parsedNumber / 10 ** 9;
-
-      if (exceedFlag) {
-        return Messages.alerts.resourcesCapacityExceeding(
-          fieldLabel,
-          processedValue,
-          units
-        );
-      }
-      return Messages.labels.estimated(processedValue, units);
-    }
-    return '';
-  };
 
   const resourceSizePerNode: ResourceSize = watch(
     DbWizardFormFields.resourceSizePerNode
